+++ src/main/java/lotto/domain/LottoNumber.java
@@ -6,28 +6,28 @@ import lotto.view.exception.LottoInputException;
 import lotto.view.message.LottoInputExceptionMessage;
 
 public class LottoNumber {
-    protected long number;
+    protected int number;
 
-    protected LottoNumber(final long number) {
+    protected LottoNumber(final int number) {
         this.number = number;
     }
 
-    public static LottoNumber from(final long number) {
+    public static LottoNumber from(final int number) {
         validate(number);
         return new LottoNumber(number);
     }
 
-    protected static void validate(final long number) {
+    protected static void validate(final int number) {
         isInAppropriateRange(number);
     }
 
-    protected static void isInAppropriateRange(final long number) {
+    protected static void isInAppropriateRange(final int number) {
         if (isNotInAppropriateRange(number)) {
             throw LottoInputException.of(LottoInputExceptionMessage.LOTTO_NUMBER_IS_NOT_IN_APPROPRIATE_RANGE);
         }
     }
 
-    protected static boolean isNotInAppropriateRange(final long number) {
+    protected static boolean isNotInAppropriateRange(final int number) {
         return number < LottoConstant.LOTTO_MIN_NUM.getValue() || number > LottoConstant.LOTTO_MAX_NUM.getValue();
     }
